Move test files from src/test to tests/
